src/requirements.txt
python-dateutil==2.8.2
pyyaml==6.0.1
pymongo==4.5.0